EDM-1039: Make CLI code windows compatible
