+++ internal/types/agent.go
@@ -1,0 +1,26 @@
+package types
+
+type EnrollmentService struct {
+	Config
+
+	// EnrollmentUIEndpoint is the address of the device enrollment UI
+	EnrollmentUIEndpoint string `json:"enrollment-ui-endpoint,omitempty"`
+}
+
+type ManagementService struct {
+	Config
+}
+
+func (s *EnrollmentService) Equal(s2 *EnrollmentService) bool {
+	if s == s2 {
+		return true
+	}
+	return s.Config.Equal(&s2.Config) && s.EnrollmentUIEndpoint == s2.EnrollmentUIEndpoint
+}
+
+func (s *ManagementService) Equal(s2 *ManagementService) bool {
+	if s == s2 {
+		return true
+	}
+	return s.Config.Equal(&s2.Config)
+}
